requirements.txt
astroid==2.5.2
colorama==0.4.4
cycler==0.10.0
decorator==4.4.2
imageio==2.9.0
isort==5.8.0
kiwisolver==1.3.1
lazy-object-proxy==1.6.0
matplotlib==3.4.2
mccabe==0.6.1
networkx==2.5.1
numpy==1.20.3
opencv-python==4.5.2.52
Pillow==8.2.0
pylint==2.7.4
pyparsing==2.4.7
PySide6==6.0.3
python-dateutil==2.8.1
PyWavelets==1.1.1
scikit-image==0.18.1
scipy==1.6.3
shiboken6==6.0.3
six==1.16.0
tifffile==2021.4.8
toml==0.10.2
wrapt==1.12.1
